+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.6
 
 - Support attribute unidirectional/bidirectional data-binding
+- Support `select` element (it was blocked by a built-in Ruby method)
 - Remove pure-struct gem dependency as the latest Opal fixed the implementation of Struct
 
 ## 0.0.5